flask/requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
Flask-SQLAlchemy